// packages/engine/src/particles/functions/ParticleEmitterMesh.d.ts
import { Mesh, Vector3, Color, Texture } from "three";
export declare class ParticleEmitterMesh extends Mesh {
    initialPositions: number[];
    initialAges: number[];
    startSize: number;
    endSize: number;
    sizeRandomness: number;
    startVelocity: Vector3;
    endVelocity: Vector3;
    angularVelocity: number;
    particleCount: number;
    lifetime: number;
    lifetimes: number[];
    lifetimeRandomness: number;
    particleSizeRandomness: number[];
    ageRandomness: number;
    ages: number[];
    colors: number[];
    endColor: Color;
    middleColor: Color;
    startColor: Color;
    startOpacity: number;
    middleOpacity: number;
    endOpacity: number;
    colorCurve: string;
    velocityCurve: string;
    sizeCurve: string;
    worldScale: Vector3;
    inverseWorldScale: Vector3;
    count: number;
    constructor(texture: Texture, configs: any);
    updateParticles(): void;
    update(dt: number): void;
    copy(source: this, recursive?: boolean): this;
    static fromArgs(args: any): Promise<ParticleEmitterMesh>;
}